templates/copilot-agent-ms-graph/requirements.txt
requests
msgraph-sdk
azure-identity
aiohttp
asyncio
configparser
rich
python-dotenv
microsoft-kiota-abstractions
microsoft-kiota-authentication-azure
requests_toolbelt
semantic-kernel
chainlit
tiktoken
openai
pillow
tenacity
json-repair
pandas
python-socketio
uvicorn
fastapi
pydantic==2.9.2
teams-ai>=1.6.0,<2.0.0